rename openrag_eval to openrag_workbench
