--- evaluation/openrag_eval/src/openrag_eval/__init__.py
@@ -1,6 +1,0 @@
-"""OpenRAG Evaluation Tool."""
-
-from openrag_eval import boards
-
-__version__ = "0.1.0"
-__all__ = ["boards"]
