--- evaluation/openrag_eval/tests/__init__.py
@@ -1,1 +1,0 @@
-"""Tests for openrag_eval."""
