+++ evaluation/openrag_workbench/src/openrag_workbench/__init__.py
@@ -1,0 +1,6 @@
+"""OpenRAG Workbench Tool."""
+
+from openrag_workbench import boards
+
+__version__ = "0.1.0"
+__all__ = ["boards"]
