+++ evaluation/openrag_workbench/src/openrag_workbench/boards/table_rich/board.yaml
