+++ evaluation/openrag_workbench/src/openrag_workbench/logging_config.py
@@ -1,4 +1,4 @@
-"""Shared logging configuration for openrag_eval."""
+"""Shared logging configuration for openrag_workbench."""
 
 import logging
