+++ evaluation/openrag_workbench/tests/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for openrag_workbench."""
